Add test for myCreateElement function of helpers.js in views - Move test folders into src folder
