+++ src/test/models/project.test.js
@@ -1,4 +1,4 @@
-import Project from '../../src/scripts/modules/models/project';
+import Project from '../../scripts/modules/models/project';
 
 test('should return project object with the given arguments including title and description', () => {
   const project = new Project('My Project', 'My description');
